Add files via upload
This project looks at the AWS Honeypot data for 2013, and explores several cybersecurity questions of interest:
1. Who are the major cyber-threats in 2013?
2. Is there a relationship between source port and attack port?
3. Where is the "center of gravity" of all cyber-attacks worldwide?

This project uses R and ggplot to develop key visualization needs.  The code is embedded in the presentation file and offers several option to the user. Further, it also looks at ggplot visualization options to improve its appearance for the user and customer.

The file concludes with Lessons Learned for the beginning R programmer.  Enjoy!